refactor: modularize knowledge graph server
