+++ src/utils/ids.ts
@@ -1,0 +1,3 @@
+export function generateSessionId(): string {
+  return `proj_${Date.now()}_${Math.random().toString(36).substring(2, 15)}`;
+}
